Moved to flask extension and app factory
