+++ nummus/controllers/auth.py
@@ -87,7 +87,7 @@ def page_login() -> str | werkzeug.Response:
     if flask_login.current_user.is_authenticated:
         # If already authenticated, skip login page
         if next_url is None:
-            return flask.redirect(flask.url_for("dashboard.page"))
+            return flask.redirect(flask.url_for("common.page_dashboard"))
         return flask.redirect(next_url)
     return flask.render_template(
         "auth/login.jinja",
@@ -132,7 +132,7 @@ def login() -> str | werkzeug.Response:
 
         next_url = form.get("next")
         if next_url is None:
-            return flask.redirect(flask.url_for("dashboard.page"))
+            return flask.redirect(flask.url_for("common.page_dashboard"))
         return flask.redirect(next_url)
