+++ nummus/templates/shared/nav-drawer.jinja
@@ -5,8 +5,8 @@
   onclick="nav.closeDrawer(event)"
 >
   <a
-    href="{{ url_for('dashboard.page') }}"
-    hx-get="{{ url_for('dashboard.page') }}"
+    href="{{ url_for('common.page_dashboard') }}"
+    hx-get="{{ url_for('common.page_dashboard') }}"
     hx-disabled-elt="this"
     >nummus</a
   >
