+++ nummus/templates/shared/nav-header.jinja
@@ -1,9 +1,9 @@
 <header id="nav-header" hx-push-url="true" hx-swap="innerHTML show:window:top">
   <icon class="nav-opener" onclick="nav.openDrawer()">menu</icon>
   <a
-    href="{{ url_for('dashboard.page') }}"
+    href="{{ url_for('common.page_dashboard') }}"
     hx-disabled-elt="this"
-    hx-get="{{ url_for('dashboard.page') }}"
+    hx-get="{{ url_for('common.page_dashboard') }}"
     >nummus</a
   >
 </header>
