+++ nummus/templates/shared/nav-rail.jinja
@@ -1,7 +1,7 @@
 <nav id="nav-rail" hx-push-url="true" hx-swap="innerHTML show:window:top">
   <a
-    href="{{ url_for('dashboard.page') }}"
-    hx-get="{{ url_for('dashboard.page') }}"
+    href="{{ url_for('common.page_dashboard') }}"
+    hx-get="{{ url_for('common.page_dashboard') }}"
     hx-disabled-elt="this"
     >nummus</a
   >
